api/requirements.txt
fastapi
uvicorn
discord.py
asyncpg